requirements.txt
Django>=3.1.7,<3.2.0
flake8>=3.8.4,<3.9.0
factory_boy>=3.2.0,<3.3.0